Extract decompiler related analyses into a separate module
Note: these analyses are not registered by default in
analyses/__init.py__, because it would cause a circular
dependency with ailment:

analyses/__init__.py
 analyses/decompiler/__init__.py
  analyses/decompiler/clinic.py
   ailment/__init__.py
    ailment/analyses/__init__.py
     ailment/analyses/simplifier.py
      analyses/__init__.py
