+++ angr/analyses/__init__.py
@@ -22,12 +22,9 @@ from .static_hooker import StaticHooker
 from .reassembler import Reassembler
 from .binary_optimizer import BinaryOptimizer
 from .disassembly import Disassembly
-from .variable_recovery import VariableRecovery
+from .variable_recovery import VariableRecovery, VariableRecoveryFast
 from .identifier import Identifier
 from .callee_cleanup_finder import CalleeCleanupFinder
 from .reaching_definitions import ReachingDefinitionAnalysis
 from .calling_convention import CallingConventionAnalysis
-from .region_identifier import RegionIdentifier
-from .structurer import Structurer
-from .structured_codegen import StructuredCodeGenerator
-# from .clinic import Clinic
+
