+++ tests/test_clinic.py
@@ -4,7 +4,7 @@ import os
 import nose.tools
 
 import angr
-from angr.analyses.clinic import Clinic
+import angr.analyses.decompiler
 
 
 def main():
